requirements.txt
langchain
langchain-core
langsmith
python-dotenv
streamlit
